+++ spec/acceptance/socket_spec.rb
@@ -13,16 +13,22 @@ describe 'systemd class' do
         description   => 'vago Server Activation Socket',
         listen_stream => [ '6565' ],
         wantedby      => [ 'sockets.target' ],
+        accept => true,
       }
 
 
-      systemd::service { 'vago':
+      systemd::service { 'vago@':
         description    => 'vago server',
         requires       => [ 'vago.socket' ],
-        documentation  => 'man:in.tftpd',
         execstart      => [ "/bin/sleep 30" ],
         standard_input => 'socket',
         also           => [ 'vago.socket' ],
+        before         => Service['vago'],
+      }
+
+      service { 'vago':
+        ensure  => 'running',
+        require => Class['::systemd'],
       }
 
       EOF